fix(file comparison): Only change displayed value when byte value changes
This fixes a UI frame jitter caused by setting the current byte unit

Also in this commit, the logic has been refactored into its own
custom hook to clean up the Calculator component

PE-128
